Replace old job service package with V2 (rename packages)
