+++ src/jobservice/config/config.go
@@ -11,7 +11,7 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/vmware/harbor/src/jobservice_v2/utils"
+	"github.com/vmware/harbor/src/jobservice/utils"
 	yaml "gopkg.in/yaml.v2"
 )
