+++ src/jobservice/period/interface.go
@@ -2,7 +2,7 @@
 
 package period
 
-import "github.com/vmware/harbor/src/jobservice_v2/models"
+import "github.com/vmware/harbor/src/jobservice/models"
 
 //Interface defines operations the periodic scheduler should have.
 type Interface interface {
