+++ src/jobservice/pool/message_server_test.go
@@ -8,13 +8,13 @@ import (
 	"testing"
 	"time"
 
-	"github.com/vmware/harbor/src/jobservice_v2/opm"
+	"github.com/vmware/harbor/src/jobservice/opm"
 
-	"github.com/vmware/harbor/src/jobservice_v2/models"
-	"github.com/vmware/harbor/src/jobservice_v2/period"
-	"github.com/vmware/harbor/src/jobservice_v2/utils"
+	"github.com/vmware/harbor/src/jobservice/models"
+	"github.com/vmware/harbor/src/jobservice/period"
+	"github.com/vmware/harbor/src/jobservice/utils"
 
-	"github.com/vmware/harbor/src/jobservice_v2/tests"
+	"github.com/vmware/harbor/src/jobservice/tests"
 )
 
 var redisPool = tests.GiveMeRedisPool()
